+++ ProjectManagement.Domain/Domain/Models/EStatus.cs
@@ -1,0 +1,11 @@
+﻿namespace ProjectManagement.Domain.Models
+{
+    public enum EStatus
+    {
+        NotStarted = 1,
+        InProgress = 2,
+        OnHold = 3,
+        Completed = 4,
+        Terminated = 5
+    }
+}
