+++ ProjectManagement.Domain/Domain/Models/Task.cs
@@ -1,0 +1,8 @@
+﻿namespace ProjectManagement.Domain.Models
+{
+    public class Task : Activity
+    {
+        public long? ProjectId { get; set; }
+        public virtual Project Project { get; set; }
+    }
+}
