+++ ProjectManagement.Domain/Domain/Models/User.cs
@@ -1,0 +1,10 @@
+﻿namespace ProjectManagement.Domain.Models
+{
+    public class User
+    {
+        public long Id { get; set; }
+        public string Name { get; set; }
+        public string Department { get; set; }
+        public string Role { get; set; }
+    }
+}
