+++ ProjectManagement.Domain/Domain/Repositories/IProjectRepository.cs
@@ -1,0 +1,15 @@
+﻿using ProjectManagement.Domain.Models;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace ProjectManagement.Domain.Repositories
+{
+    public interface IProjectRepository
+    {
+        Task<IEnumerable<Project>> GetProjects();
+        Task<Project> GetProject(long id);
+        Task<bool> CreateProject(Project project);
+        Task<bool> UpdateProject(long id, Project project);
+        Task<bool> DeleteProject(long id);
+    }
+}
