+++ ProjectManagement.Domain/Domain/Repositories/IUserRepository.cs
@@ -1,0 +1,15 @@
+﻿using ProjectManagement.Domain.Models;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace ProjectManagement.Domain.Repositories
+{
+    public interface IUserRepository
+    {
+        Task<IEnumerable<User>> GetUsers();
+        Task<User> GetUser(long id);
+        Task<bool> CreateUser(User user);
+        Task<bool> UpdateUser(long id, User user);
+        Task<bool> DeleteUser(long id);
+    }
+}
