--- ProjectManagement.Domain/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace ProjectManagement.Domain
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-        }
-    }
-}
