+++ ProjectManagement.Infrastructure/Persistance/Contexts/AppDBContext.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.EntityFrameworkCore;
+using ProjectManagement.Domain.Models;
+
+namespace ProjectManagement.Infrastructure.Persistance.Contexts
+{
+    public class AppDbContext : DbContext
+    {
+        public AppDbContext(DbContextOptions<AppDbContext> options) : base(options){}
+        public virtual DbSet<Project> Projects { get; set; }
+        public virtual DbSet<Task> Tasks { get; set; }
+        public virtual DbSet<User> Users { get; set; }
+    }
+}
