+++ ProjectManagement.Infrastructure/Persistance/Repositories/BaseRepository.cs
@@ -1,0 +1,14 @@
+﻿using ProjectManagement.Infrastructure.Persistance.Contexts;
+
+namespace ProjectManagement.Infrastructure.Persistance.Repositories
+{
+    public abstract class BaseRepository
+    {
+        protected readonly AppDbContext _context;
+
+        public BaseRepository(AppDbContext context)
+        {
+            _context = context;
+        }
+    }
+}
