--- ProjectManagementAPI/Domain/Models/EStatus.cs
@@ -1,11 +1,0 @@
-﻿namespace ProjectManagementAPI.Models
-{
-    public enum EStatus
-    {
-        NotStarted = 1,
-        InProgress = 2,
-        OnHold = 3,
-        Completed = 4,
-        Terminated = 5
-    }
-}
