--- ProjectManagementAPI/Domain/Models/Project.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace ProjectManagementAPI.Models
-{
-    public class Project : Activity
-    {
-        public virtual ICollection<Task> Tasks { get; set; } = new List<Task>();
-    }
-}
