--- ProjectManagementAPI/Domain/Models/Task.cs
@@ -1,8 +1,0 @@
-﻿namespace ProjectManagementAPI.Models
-{
-    public class Task : Activity
-    {
-        public long? ProjectId { get; set; }
-        public virtual Project Project { get; set; }
-    }
-}
