--- ProjectManagementAPI/Domain/Repositories/ITaskRepository.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace ProjectManagementAPI.Domain.Repositories
-{
-    public interface ITaskRepository
-    {
-        Task<IEnumerable<Models.Task>> GetTasks();
-        Task<Models.Task> GetTask(long id);
-        Task<bool> CreateTask(Models.Task task);
-        Task<bool> UpdateTask(long id, Models.Task task);
-        Task<bool> DeleteTask(long id);
-    }
-}
