--- ProjectManagementAPI/Domain/Services/IUserService.cs
@@ -1,15 +1,0 @@
-﻿using ProjectManagementAPI.Models;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace ProjectManagementAPI.Interfaces
-{
-    public interface IUserService
-    {
-        Task<IEnumerable<User>> GetUsers();
-        Task<User> GetUser(long id);
-        Task<bool> CreateUser(User user);
-        Task<bool> UpdateUser(long id, User user);
-        Task<bool> DeleteUser(long id);
-    }
-}
