--- ProjectManagementAPI/Persistance/Contexts/AppDBContext.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace ProjectManagementAPI.Models
-{
-    public class AppDbContext : DbContext
-    {
-        public AppDbContext(DbContextOptions<AppDbContext> options) : base(options){}
-        public virtual DbSet<Project> Projects { get; set; }
-        public virtual DbSet<Task> Tasks { get; set; }
-        public virtual DbSet<User> Users { get; set; }
-    }
-}
