--- ProjectManagementAPI/Persistance/Repositories/BaseRepository.cs
@@ -1,14 +1,0 @@
-﻿using ProjectManagementAPI.Models;
-
-namespace ProjectManagementAPI.Persistance.Repositories
-{
-    public abstract class BaseRepository
-    {
-        protected readonly AppDbContext _context;
-
-        public BaseRepository(AppDbContext context)
-        {
-            _context = context;
-        }
-    }
-}
