--- ProjectManagementAPI/Persistance/Repositories/UserRepository.cs
@@ -1,52 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using ProjectManagementAPI.Domain.Repositories;
-using ProjectManagementAPI.Models;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace ProjectManagementAPI.Persistance.Repositories
-{
-    public class UserRepository : BaseRepository, IUserRepository
-    {
-        public UserRepository(AppDbContext context) : base(context) { }
-
-        public async Task<bool> CreateUser(User user)
-        {
-            _context.Users.Add(user);
-            await _context.SaveChangesAsync();
-            return true;
-        }
-
-        public async Task<bool> DeleteUser(long id)
-        {
-            var user = _context.Users.Find(id);
-            if (user == null)
-            {
-                return false;
-            }
-            _context.Users.Remove(user);
-            await _context.SaveChangesAsync();
-            return true;
-        }
-
-        public async Task<User> GetUser(long id)
-        {
-            var user = await _context.Users.FirstOrDefaultAsync(u => u.Id == id);
-            return user;
-        }
-
-        public async Task<IEnumerable<User>> GetUsers() => await _context.Users.ToListAsync();
-
-        public async Task<bool> UpdateUser(long id, User user)
-        {
-            var foundUser = await _context.Users.FindAsync(id);
-            if (foundUser == null)
-            {
-                return false;
-            }
-            _context.Entry(user).State = EntityState.Modified;
-            await _context.SaveChangesAsync();
-            return true;
-        }
-    }
-}
