+++ ProjectManagementAPI/ProjectManagement.API.csproj
@@ -6,50 +6,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Migrations\20190614102522_initialcreate.cs" />
-    <Compile Remove="Migrations\20190614102522_initialcreate.Designer.cs" />
-    <Compile Remove="Migrations\20190614124108_Users.cs" />
-    <Compile Remove="Migrations\20190614124108_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190614125630_Users.cs" />
-    <Compile Remove="Migrations\20190614125630_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190614133309_Users.cs" />
-    <Compile Remove="Migrations\20190614133309_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190614135010_Users.cs" />
-    <Compile Remove="Migrations\20190614135010_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190614144135_Users.cs" />
-    <Compile Remove="Migrations\20190614144135_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190614150955_Users.cs" />
-    <Compile Remove="Migrations\20190614150955_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190614153114_Users.cs" />
-    <Compile Remove="Migrations\20190614153114_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190614154213_Users.cs" />
-    <Compile Remove="Migrations\20190614154213_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190614154331_Users.cs" />
-    <Compile Remove="Migrations\20190614154331_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190614160156_Users.cs" />
-    <Compile Remove="Migrations\20190614160156_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190718141647_Users.cs" />
-    <Compile Remove="Migrations\20190718141647_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190718154757_Users.cs" />
-    <Compile Remove="Migrations\20190718154757_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190718155140_Users.cs" />
-    <Compile Remove="Migrations\20190718155140_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190718171829_Users.cs" />
-    <Compile Remove="Migrations\20190718171829_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190718172653_Users.cs" />
-    <Compile Remove="Migrations\20190718172653_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190718182031_Users.cs" />
-    <Compile Remove="Migrations\20190718182031_Users.Designer.cs" />
-    <Compile Remove="Migrations\20190718183045_Users.cs" />
-    <Compile Remove="Migrations\20190718183045_Users.Designer.cs" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc2" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.4" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer.Design" Version="1.1.6" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc2" />
+    <ProjectReference Include="..\ProjectManagement.Domain\ProjectManagement.Domain.csproj" />
+    <ProjectReference Include="..\ProjectManagement.Infrastructure\ProjectManagement.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
